Create CDM/CIM Views
More refactoring of the CDM/CIM implmentation. Created project WFS and WFSTest to hold WFSx classes.
Moved code out of the CDMTable abd CIMTable into WFS classes.
This is functional.
