--- src/Impl/WFSCDMCIM.cs
@@ -1,15 +1,0 @@
-﻿
-using Contract;
-
-namespace Impl
-{
-   public class WFSCDMCIM
-   {
-      protected IContext ctx; 
-
-      public WFSCDMCIM(IContext ctx)
-      {
-         this.ctx = ctx;
-      }
-   }
-}
