+++ src/ImplTests/LogFindTests.cs
@@ -1,7 +1,8 @@
-﻿using Impl;
+﻿using Samples;
+using Impl;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ImplTests
+namespace SPLogParserTests
 {
    [TestClass]
    public class LogFindTests
