--- src/ImplTests/_wfs_inf_cim_cash_in_statusTests.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ImplTests
-{
-   class _wfs_inf_cim_cash_in_statusTests
-   {
-   }
-}
