--- src/ImplTests/samples_note_numbers.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ImplTests
-{
-   public static class samples_note_numbers
-   {
-   }
-}
