+++ src/Samples/SampleTests.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{1D0E5765-4FB9-437C-AD8A-7BE981229677}</ProjectGuid>
+    <ProjectGuid>{915D5FB6-69C8-4BAB-984B-090801A300CA}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>CDMView</RootNamespace>
-    <AssemblyName>CDMView</AssemblyName>
+    <RootNamespace>Samples</RootNamespace>
+    <AssemblyName>Samples</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -32,7 +32,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
-    <Reference Include="System.ComponentModel.Composition" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -42,34 +41,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CDMTable.cs" />
-    <Compile Include="CDMView.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Contract\Contract.csproj">
-      <Project>{9ef4286e-dc11-479c-84a0-eb43a70e1bde}</Project>
-      <Name>Contract</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Impl\Impl.csproj">
-      <Project>{f89403e5-4c5c-4d8f-8401-b2d56f6d8a2b}</Project>
-      <Name>Impl</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\WFS\WFS.csproj">
-      <Project>{8685a55b-44fe-4e97-a4d6-2915d537af3e}</Project>
-      <Name>WFS</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="CDMView.xsd" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="CDMView.xml" />
+    <Compile Include="samples_cdm.cs" />
+    <Compile Include="samples_cim.cs" />
+    <Compile Include="samples_lppcashin.cs" />
+    <Compile Include="samples_note_numbers.cs" />
+    <Compile Include="samples_tracefilereader.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
-  <PropertyGroup>
-    <PostBuildEvent>copy "$(MSBuildProjectDirectory)\bin\$(Configuration)\*.dll" $(SolutionDir)\..\dist
-copy "$(MSBuildProjectDirectory)\CDMView.xsd" $(SolutionDir)\..\dist
-copy "$(MSBuildProjectDirectory)\CDMView.xml" $(SolutionDir)\..\dist</PostBuildEvent>
-  </PropertyGroup>
 </Project>
