+++ src/Samples/samples_cdm.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace ImplTests
+namespace Samples
 {
    public static class samples_cdm
    {
@@ -16,15 +16,15 @@ lpResult =
 	{
 		fwDevice = [0],
 		fwSafeDoor = [1],
-		fwDispenser = [0],
-		fwIntermediateStacker = [0],
+		fwDispenser = [2],
+		fwIntermediateStacker = [3],
 		lppPositions =
 		{
 			fwPosition = [4],
-			fwShutter = [0],
-			fwPositionStatus = [0],
-			fwTransport = [0],
-			fwTransportStatus = [0]
+			fwShutter = [5],
+			fwPositionStatus = [6],
+			fwTransport = [7],
+			fwTransportStatus = [8]
 		}
 		lpszExtra = [BCMode=REAL-NOTE,BRM20SensorInfo=[1][4][5][0][0][0][0][0][2][2][2][2][2][1][1][0][0][0][0][0][0][0][0][0][0][0][0][0][0][0][0][0][3][1][1][1][1][1][1][1][4][3][3][4][3][3][0][0][3][0][0][3][3][4][0][0][6][6][6][0][0][0][3][0][1][0][0][0][0][0][0][0][...],
 		dwGuidLights =
@@ -36,9 +36,9 @@ lpResult =
 			[0x00000000], [0x00000000], [0x00000000], [0x00000000],
 			[0x00000000], [0x00000000], [0x00000000], [0x00000000],
 			[0x00000000], [0x00000000], [0x00000000], [0x00000000]
-		wDevicePosition = [0],
-		usPowerSaveRecoveryTime = [0],
-		wAntiFraudModule = [0]
+		wDevicePosition = [9],
+		usPowerSaveRecoveryTime = [10],
+		wAntiFraudModule = [11]
 	}
 }";
       public const string WFS_INF_CDM_STATUS_2 = @"29114294967295018600090735020007MONIMNG0008XFSAGENT00102023/01/24001221:46 33.7550004INFO0015Agent::FireTrap0092call WFSAsyncGetInfo(ManagedName=[CashDispenser] hService=[28] dwCategory=[301] hResult=[0])01864294967295016700090735030006COMMON0009FRAMEWORK00102023/01/24001221:46 33.7550011INFORMATION0028CServiceProvider::GetService0053service OK {HSERVICE[28], LogicalName[CashDispenser]}01674294967295020000090735040003CIM0007ACTIVEX00102023/01/24001221:46 33.7560006XFSAPI0023CContextMgr::MgrWndProc0101WFS_GETINFO_COMPLETE(RequestID=10022, hService=12, hResult=0, dwCommandCode=1303 lpBuffer=0x255C521C)02004294967295016600090735050003CDM0009FRAMEWORK00102023/01/24001221:46 33.7560011INFORMATION0021CBaseService::GetInfo0062Srvc=301 ReqID=10023 Wnd=0x00020620 Cmd=301 TimeOut=60000 IO=501664294967295015900090735060003CIM0007ACTIVEX00102023/01/24001221:46 33.7560011INFORMATION0022CMsgWnd::DefWindowProc0056message=0x00000407, wParam=0x3764E760, lParam=0x252011BC01594294967295015800090735070013CashDispenser0009FRAMEWORK00102023/01/24001221:46 33.7560007DEVCALL0021CBaseService::GetInfo0048HSERVICE[28] CATEGORY[301] IN BUFFER[0x00000000]01584294967295015500090735080003CIM0007ACTIVEX00102023/01/24001221:46 33.7560011INFORMATION0031CCimService::HandleCashUnitInfo0043GetInfo-Result[CashUnitInfo] = {hResult[0]}01554294967295017100090735090013CashDispenser0009FRAMEWORK00102023/01/24001221:46 33.7560007DEVRETN0021CBaseService::GetInfo0061HSERVICE[28] CATEGORY[301] HRESULT[0], OUT BUFFER[0x014304F4]01714294967295012400090735100006COMMON0003SPI00102023/01/24001221:46 33.7570011INFORMATION0010WFPGetInfo0034HSERVICE=12, SrvcVersion=2563(A03)01244294967295141200090735110003CDM0003SPI00102023/01/24001221:46 33.7570009XFS_EVENT0012GETINFO[301]1325WFS_GETINFO_COMPLETE, 
@@ -414,9 +414,9 @@ lpResult =
 	{
 		cCurrencyID = [USD],
 		ulAmount = [30],
-		usCount = [6],
-		lpulValues = [0, 0, 0, 2, 1, 0],
-		ulCashBox = [0]
+		usCount = [31],
+		lpulValues = [32, 33, 34, 35, 36, 37],
+		ulCashBox = [38]
 	}
 }";
       public const string WFS_CMD_CDM_DISPENSE_2 = @"02434294967295019501781740810003CDM0007ACTIVEX00102023/01/28001216:21 26.0090006XFSAPI0022CService::AsyncGetInfo0097WFSAsyncGetInfo(hService=2, dwCategory=301, lpQueryDetails=0x00000000, RequestID=16530) hResult=001954294967295013501781740820003CDM0009FRAMEWORK00102023/01/28001216:21 26.0090011INFORMATION0020CCdmService::Execute0032WFS_CMD_CDM_DISPENSE{hResult[0]}01354294967295015301781740830003CDM0009FRAMEWORK00102023/01/28001216:21 26.0090011INFORMATION0027CDM_POST_SYSE_DEVICE_STATUS0043Invoke[wDevStatus = 0, lpszPhysicalName = ]01534294967295019601781740840003CDM0007ACTIVEX00102023/01/28001216:21 26.0090006XFSAPI0022CService::AsyncGetInfo0098WFSAsyncGetInfo(hService=18, dwCategory=301, lpQueryDetails=0x00000000, RequestID=16531) hResult=001964294967295015901781740850003CDM0007ACTIVEX00102023/01/28001216:21 26.0090011INFORMATION0022CMsgWnd::DefWindowProc0056message=0x00000415, wParam=0x178318E8, lParam=0x1E77805401594294967295016801781740860013CashDispenser0009FRAMEWORK00102023/01/28001216:21 26.0090007DEVRETN0021CBaseService::Execute0058HSERVICE[2] COMMAND[302] HRESULT[0] OUT BUFFER[0x00A78B04]01684294967295015301781740870003CDM0009FRAMEWORK00102023/01/28001216:21 26.0090011INFORMATION0031CCdmService::CheckStatusChanged0039Device Status Message Report.[Status=0]01534294967295015901781740880003CDM0007ACTIVEX00102023/01/28001216:21 26.0090011INFORMATION0022CMsgWnd::DefWindowProc0056message=0x00000415, wParam=0x07BDAA50, lParam=0x094B01B401594294967295011901781740890006COMMON0003SPI00102023/01/28001216:21 26.0100011INFORMATION0010WFPGetInfo0029HSERVICE=2, SrvcVersion=3(03)01194294967295012201781740900006COMMON0009FRAMEWORK00102023/01/28001216:21 26.0100011INFORMATION0015CProcessor::Run0021---- Wait [6084] ----01224294967295042701781740910003CDM0003SPI00102023/01/28001216:21 26.0100009XFS_EVENT0012EXECUTE[302]0340WFS_EXECUTE_COMPLETE, 
@@ -589,5 +589,242 @@ lpResult =
 ";
       public const string WFS_SRVE_CDM_ITEMSTAKEN_2 = @"";
       public const string WFS_SRVE_CDM_ITEMSTAKEN_3 = @"";
+
+      public const string GET_INFO_COMPLETE_AS_TABLE =@"02424294967295016600358535370006COMMON0009FRAMEWORK00102022/12/19001216:01 26.0150011INFORMATION0028CServiceProvider::GetService0052service OK {HSERVICE[2], LogicalName[CashDispenser]}01664294967295014500358535380003BRM0002SP00102022/12/19001216:01 26.0150011INFORMATION0019CMonitorThread::Run0050m_oldBrmStatus.OutputPosition.fwHand changed(1->0)01454294967295015900358535390003CDM0007ACTIVEX00102022/12/19001216:01 26.0150011INFORMATION0022CMsgWnd::DefWindowProc0056message=0x00000407, wParam=0x01493AD8, lParam=0x08F07C5C01594294967295019500358535400003CDM0007ACTIVEX00102022/12/19001216:01 26.0150006XFSAPI0022CService::AsyncGetInfo0097WFSAsyncGetInfo(hService=2, dwCategory=301, lpQueryDetails=0x00000000, RequestID=39299) hResult=001954294967295016700358535410003CDM0009FRAMEWORK00102022/12/19001216:01 26.0160011INFORMATION0021CBaseService::GetInfo0063Srvc=303 ReqID=39297 Wnd=0x00010580 Cmd=303 TimeOut=300000 IO=101674294967295015600358535420003CDM0007ACTIVEX00102022/12/19001216:01 26.0160011INFORMATION0025CCdmService::HandleStatus0050GetInfo-Result[Status][ReqID=39296] = {hResult[0]}01564294967295011900358535430003BRM0002SP00102022/12/19001216:01 26.0160011INFORMATION0019CMonitorThread::Run0024fwChecking changed(0->1)01194294967295015700358535440013CashDispenser0009FRAMEWORK00102022/12/19001216:01 26.0160007DEVCALL0021CBaseService::GetInfo0047HSERVICE[2] CATEGORY[303] IN BUFFER[0x00000000]01574294967295015800358535450003CDM0007ACTIVEX00102022/12/19001216:01 26.0160011INFORMATION0025CCdmService::HandleStatus0052StStacker-old, StStacker-new : { NOTEMPTY , EMPTY }.01584294967295015900358535460003CDM0007ACTIVEX00102022/12/19001216:01 26.0160011INFORMATION0022CMsgWnd::DefWindowProc0056message=0x00000407, wParam=0x28B3B7A0, lParam=0x392EF0B401594294967295015700358535470003CDM0007ACTIVEX00102022/12/19001216:01 26.0170005EVENT0041CCdmService::SetStackerStatusChangedEvent0041FireStackerStatusChanged{NewValue[EMPTY]}01574294967295015400358535480003CDM0009FRAMEWORK00102022/12/19001216:01 26.0170011INFORMATION0032CBaseService::CheckStatusChanged0039Device Status Message Report.[Status=0]01544294967295015600358535490003CDM0007ACTIVEX00102022/12/19001216:01 26.0170011INFORMATION0025CCdmService::HandleStatus0050GetInfo-Result[Status][ReqID=39294] = {hResult[0]}01564294967295017000358535500013CashDispenser0009FRAMEWORK00102022/12/19001216:01 26.0170007DEVRETN0021CBaseService::GetInfo0060HSERVICE[2] CATEGORY[303] HRESULT[0], OUT BUFFER[0x03105C6C]01704294967295015500358535510003CDM0007ACTIVEX00102022/12/19001216:01 26.0170011INFORMATION0025CCdmService::HandleStatus0049StShutter-old, StShutter-new : { CLOSED , OPEN }.01554294967295015900358535520003CDM0009FRAMEWORK00102022/12/19001216:01 26.0170011INFORMATION0028CBaseService::BroadcastEvent0048hApp=0x00000000, pCursor->pData->hApp=0x118D5E7801594294967295015800358535530003CDM0007ACTIVEX00102022/12/19001216:01 26.0170011INFORMATION0025CCdmService::HandleStatus0052StStacker-old, StStacker-new : { NOTEMPTY , EMPTY }.01584294967295022300358535540006COMMON0009FRAMEWORK00102022/12/19001216:01 26.0180011INFORMATION0033CSpCmdDispatcher::DispatchCommand0104pService->GetInfo() {HSERVICE[2], HWND[0x00010580], REQUESTID[39298], dwCmdCode[309], dwTimeOut[300000]}02234294967295015600358535550003CDM0007ACTIVEX00102022/12/19001216:01 26.0180005EVENT0041CCdmService::SetShutterStatusChangedEvent0040FireShutterStatusChanged{NewValue[OPEN]}01564294967295160600358535560003CDM0003SPI00102022/12/19001216:01 26.0180009XFS_EVENT0012GETINFO[303]1519WFS_GETINFO_COMPLETE, 
+lpResult =
+{
+	hWnd = [0x00010580],
+	RequestID = [39297],
+	hService = [2],
+	tsTimestamp = [2022/12/19 16:01 26.018],
+	hResult = [0],
+	u.dwCommandCode = [303],
+	lpBuffer = [0x526afb84]
+	{
+		usTellerID=0
+		usCount=6
+		lppList->
+		{
+			usNumber		1		2		3		4		5		6
+			usType			6		7		8		9		10		11
+			lpszCashUnitName	[LCU00]		[LCU01]		[LCU02]		[LCU03]		[LCU04]		[LCU05]		
+			cUnitID			LCU00		LCU01		LCU02		LCU03		LCU04		LCU05
+			cCurrencyID		   		   		USA		USB		USC		USD
+			ulValues		12		13		14		15		16		17
+			ulInitialCount		18		19		20		21		22		23
+			ulCount			24		25		26		27		28		29
+			ulRejectCount		30		31		32		33		34		35
+			ulMinimum		36		37		38		39		40		41
+			ulMaximum		42		43		44		45		46		47
+			bAppLock		0		0		0		0		0		0
+			usStatus		48		49		50		51		52		53
+			ulDispensedCount	54		55		56		57		58		59		
+			ulPresentedCount	60		61		62		63		64		65		
+			ulRetractedCount	66		67		68		69		70		71		
+
+			lppPhysical->
+			usNumPhysicalCUs	1		1		1		1		1		1		
+			lpPhysicalPositionName	RetractCassette	RejectCassette	CassetteA	CassetteB	CassetteC	CassetteD	
+			cUnitID			RTCST		RJCST		CST_A		CST_B		CST_C		CST_D
+			ulInitialCount		0		0		1400		1400		1400		1400
+			ulCount			0		2		1364		1354		1377		1352
+			ulRejectCount		0		2		1		0		1		0
+			ulMaximum		80		210		0		0		0		0
+			usPStatus		4		0		0		0		0		0
+			bHardwareSensor		1		1		1		1		1		1
+			ulDispensedCount	0		0		71		60		154		67		
+			ulPresentedCount	0		0		65		57		146		67		
+			ulRetractedCount	0		0		0		0		0		0		
+
+		}
+	}
+}";
+      public const string GET_INFO_COMPLETE_AS_LIST = @"
+14144294967295016600159585730006COMMON0009FRAMEWORK00102022/12/07001210:18 59.5080011INFORMATION0028CServiceProvider::GetService0052service OK {HSERVICE[2], LogicalName[CashDispenser]}01664294967295015900159585740003CDM0007ACTIVEX00102022/12/07001210:18 59.5080011INFORMATION0022CMsgWnd::DefWindowProc0056message=0x00000407, wParam=0x1DE33088, lParam=0x1BFBBB7401594294967295016700159585750003CDM0009FRAMEWORK00102022/12/07001210:18 59.5080011INFORMATION0021CBaseService::GetInfo0063Srvc=303 ReqID=59758 Wnd=0x00020108 Cmd=303 TimeOut=300000 IO=101674294967295015600159585760003CDM0007ACTIVEX00102022/12/07001210:18 59.5090011INFORMATION0025CCdmService::HandleStatus0050GetInfo-Result[Status][ReqID=59757] = {hResult[0]}01564294967295015700159585770013CashDispenser0009FRAMEWORK00102022/12/07001210:18 59.5090007DEVCALL0021CBaseService::GetInfo0047HSERVICE[2] CATEGORY[303] IN BUFFER[0x00000000]01574294967295015500159585780003CDM0007ACTIVEX00102022/12/07001210:18 59.5090011INFORMATION0025CCdmService::HandleStatus0049StShutter-old, StShutter-new : { OPEN , CLOSED }.01554294967295015800159585790003CDM0007ACTIVEX00102022/12/07001210:18 59.5090005EVENT0041CCdmService::SetShutterStatusChangedEvent0042FireShutterStatusChanged{NewValue[CLOSED]}01584294967295012200159585800003CDM0007ACTIVEX00102022/12/07001210:18 59.5110008PROPERTY0021Ctrl::GetDeviceStatus0023DeviceStatus[DEVONLINE]01224294967295019300159585810003CDM0007ACTIVEX00102022/12/07001210:18 59.5120011INFORMATION0024CContextMgr::SetComplete0088SetComplete(Waiting ReqID=[59757], Arrival ReqID=[59757]), Remove ReqID[59757] from List01934294967295011800159585820003CDM0007ACTIVEX00102022/12/07001210:18 59.5130005EVENT0029CCdmService::HandleItemsTaken0014FireItemsTaken01184294967295014100159585830003CDM0009FRAMEWORK00102022/12/07001210:18 59.5140011INFORMATION0020CCdmService::GetInfo0038WFS_INF_CDM_CASH_UNIT_INFO{hResult[0]}01414294967295017000159585840013CashDispenser0009FRAMEWORK00102022/12/07001210:18 59.5140007DEVRETN0021CBaseService::GetInfo0060HSERVICE[2] CATEGORY[303] HRESULT[0], OUT BUFFER[0x00B24B4C]01704294967295022300159585850006COMMON0009FRAMEWORK00102022/12/07001210:18 59.5140011INFORMATION0033CSpCmdDispatcher::DispatchCommand0104pService->GetInfo() {HSERVICE[2], HWND[0x00020108], REQUESTID[59759], dwCmdCode[301], dwTimeOut[300000]}02234294967295016600159585860006COMMON0009FRAMEWORK00102022/12/07001210:18 59.5150011INFORMATION0028CServiceProvider::GetService0052service OK {HSERVICE[2], LogicalName[CashDispenser]}01664294967295418200159585870003CDM0003SPI00102022/12/07001210:18 59.5150009XFS_EVENT0012GETINFO[303]4095WFS_GETINFO_COMPLETE, 
+lpResult =
+{
+	hWnd = [0x00020108],
+	RequestID = [59758],
+	hService = [2],
+	tsTimestamp = [2022/12/07 10:18 59.514],
+	hResult = [0],
+	u.dwCommandCode = [303],
+	lpBuffer = [0x20f11d6c]
+	{
+		usTellerID = [0],
+		usCount = [5],
+		lppList =
+		{
+			usNumber = [1],
+			usType = [11],
+			lpszCashUnitName = NULL,
+			cUnitID = [LCU00],
+			cCurrencyID = [   ],
+			ulValues = [12],
+			ulInitialCount = [13],
+			ulCount = [14],
+			ulRejectCount = [15],
+			ulMinimum = [16],
+			ulMaximum = [17],
+			bAppLock = [0],
+			usStatus = [18],
+			usNumPhysicalCUs = [1],
+			lppPhysical = 
+			{
+				lpPhysicalPositionName = [Cst_0],
+				cUnitID = [PCU00],
+				ulInitialCount = [0],
+				ulCount = [0],
+				ulRejectCount = [0],
+				ulMaximum = [300],
+				usPStatus = [0],
+				bHardwareSensor = [0],
+				ulDispensedCount = [19],
+				ulPresentedCount = [110],
+				ulRetractedCount = [111]
+			}
+			ulDispensedCount = [19],
+			ulPresentedCount = [110],
+			ulRetractedCount = [111]
+		}
+		{
+			usNumber = [2],
+			usType = [21],
+			lpszCashUnitName = NULL,
+			cUnitID = [LCU01],
+			cCurrencyID = [USD],
+			ulValues = [22],
+			ulInitialCount = [23],
+			ulCount = [24],
+			ulRejectCount = [25],
+			ulMinimum = [26],
+			ulMaximum = [27],
+			bAppLock = [0],
+			usStatus = [28],
+			usNumPhysicalCUs = [1],
+			lppPhysical = 
+			{
+				lpPhysicalPositionName = [Cst_1],
+				cUnitID = [PCU01],
+				ulInitialCount = [3000],
+				ulCount = [3000],
+				ulRejectCount = [0],
+				ulMaximum = [0],
+				usPStatus = [0],
+				bHardwareSensor = [0],
+				ulDispensedCount = [29],
+				ulPresentedCount = [210],
+				ulRetractedCount = [211]
+			}
+			ulDispensedCount = [29],
+			ulPresentedCount = [210],
+			ulRetractedCount = [211]
+		}
+		{
+			usNumber = [3],
+			usType = [31],
+			lpszCashUnitName = NULL,
+			cUnitID = [LCU02],
+			cCurrencyID = [ABC],
+			ulValues = [32],
+			ulInitialCount = [33],
+			ulCount = [34],
+			ulRejectCount = [35],
+			ulMinimum = [36],
+			ulMaximum = [37],
+			bAppLock = [0],
+			usStatus = [38],
+			usNumPhysicalCUs = [1],
+			lppPhysical = 
+			{
+				lpPhysicalPositionName = [Cst_2],
+				cUnitID = [PCU02],
+				ulInitialCount = [3000],
+				ulCount = [3000],
+				ulRejectCount = [0],
+				ulMaximum = [0],
+				usPStatus = [0],
+				bHardwareSensor = [1],
+				ulDispensedCount = [39],
+				ulPresentedCount = [310],
+				ulRetractedCount = [311]
+			}
+			ulDispensedCount = [39],
+			ulPresentedCount = [310],
+			ulRetractedCount = [311]
+		}
+		{
+			usNumber = [4],
+			usType = [41],
+			lpszCashUnitName = NULL,
+			cUnitID = [LCU03],
+			cCurrencyID = [CAD],
+			ulValues = [42],
+			ulInitialCount = [43],
+			ulCount = [44],
+			ulRejectCount = [45],
+			ulMinimum = [46],
+			ulMaximum = [47],
+			bAppLock = [0],
+			usStatus = [48],
+			usNumPhysicalCUs = [1],
+			lppPhysical = 
+			{
+				lpPhysicalPositionName = [Cst_3],
+				cUnitID = [PCU03],
+				ulInitialCount = [3000],
+				ulCount = [3000],
+				ulRejectCount = [0],
+				ulMaximum = [0],
+				usPStatus = [3],
+				bHardwareSensor = [1],
+				ulDispensedCount = [49],
+				ulPresentedCount = [410],
+				ulRetractedCount = [411]
+			}
+			ulDispensedCount = [49],
+			ulPresentedCount = [410],
+			ulRetractedCount = [411]
+		}
+		{
+			usNumber = [5],
+			usType = [51],
+			lpszCashUnitName = NULL,
+			cUnitID = [LCU04],
+			cCurrencyID = [FRA],
+			ulValues = [52],
+			ulInitialCount = [53],
+			ulCount = [54],
+			ulRejectCount = [55],
+			ulMinimum = [56],
+			ulMaximum = [57],
+			bAppLock = [0],
+			usStatus = [58],
+			usNumPhysicalCUs = [2],
+			lppPhysical = 
+			{
+				lpPhysicalPositionName = [Cst_4],
+				cUnitID = [PCU04],
+				ulInitialCount = [3000],
+				ulCount = [3000],
+				ulRejectCount = [0],
+				ulMaximum = [0],
+				usPStatus = [4],
+				bHardwareSensor = [1],
+				ulDispensedCount = [59],
+				ulPresentedCount = [510],
+				ulRetractedCount = [511]
+			}
+			{
+				lpPhysicalPositionName = [Cst_5],
+				cUnitID = [PCU05],
+				ulInitialCount = [3000],
+				ulCount = [3000],
+				ulRejectCount = [0],
+				ulMax
+		......(More Data)......
+	}
+}
+";
    }
 }
