+++ src/Samples/samples_note_numbers.cs
@@ -1,0 +1,6 @@
+﻿namespace Samples
+{
+   public static class samples_note_numbers
+   {
+   }
+}
