+++ src/Samples/samples_tracefilereader.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ImplTests
+﻿namespace Samples
 {
    public static class samples_tracefilereader
    {
