+++ src/WFS/WFS.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F89403E5-4C5C-4D8F-8401-B2D56F6D8A2B}</ProjectGuid>
+    <ProjectGuid>{8685A55B-44FE-4E97-A4D6-2915D537AF3E}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Impl</RootNamespace>
-    <AssemblyName>Impl</AssemblyName>
+    <RootNamespace>WFS</RootNamespace>
+    <AssemblyName>WFS</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -33,8 +33,6 @@
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Drawing" />
-    <Reference Include="System.IO.Compression.FileSystem" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -43,55 +41,27 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="_datatable_ops.cs" />
-    <Compile Include="BaseTable.cs" />
-    <Compile Include="BaseView.cs" />
-    <Compile Include="Context.cs" />
-    <Compile Include="FileSystemProvider.cs" />
-    <Compile Include="IdentifyLines.cs" />
-    <Compile Include="LogFind.cs" />
-    <Compile Include="Logger.cs" />
-    <Compile Include="LogTime.cs" />
-    <Compile Include="lpResult.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="LogLine.cs" />
-    <Compile Include="TraceFileReader.cs" />
-    <Compile Include="Utilities.cs" />
+    <Compile Include="WFS.cs" />
+    <Compile Include="WFSCDMCUINFO.cs" />
+    <Compile Include="WFSCDMDENOMINATION.cs" />
+    <Compile Include="WFSCDMSTATUS.cs" />
+    <Compile Include="WFSCIMCASHINFO.cs" />
+    <Compile Include="WFSCIMCASHINSTATUS.cs" />
+    <Compile Include="WFSCIMNOTENUMBERLIST.cs" />
+    <Compile Include="WFSCIMSTATUS.cs" />
+    <Compile Include="_wfs_cmd_cim_cash_in.cs" />
+    <Compile Include="_wfs_inf_cim_cash_in_status.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Contract\Contract.csproj">
       <Project>{9ef4286e-dc11-479c-84a0-eb43a70e1bde}</Project>
       <Name>Contract</Name>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <COMReference Include="Microsoft.Office.Core">
-      <Guid>{2DF8D04C-5BFA-101B-BDE5-00AA0044DE52}</Guid>
-      <VersionMajor>2</VersionMajor>
-      <VersionMinor>8</VersionMinor>
-      <Lcid>0</Lcid>
-      <WrapperTool>primary</WrapperTool>
-      <Isolated>False</Isolated>
-      <EmbedInteropTypes>True</EmbedInteropTypes>
-    </COMReference>
-    <COMReference Include="Microsoft.Office.Interop.Excel">
-      <Guid>{00020813-0000-0000-C000-000000000046}</Guid>
-      <VersionMajor>1</VersionMajor>
-      <VersionMinor>9</VersionMinor>
-      <Lcid>0</Lcid>
-      <WrapperTool>primary</WrapperTool>
-      <Isolated>False</Isolated>
-      <EmbedInteropTypes>True</EmbedInteropTypes>
-    </COMReference>
-    <COMReference Include="VBIDE">
-      <Guid>{0002E157-0000-0000-C000-000000000046}</Guid>
-      <VersionMajor>5</VersionMajor>
-      <VersionMinor>3</VersionMinor>
-      <Lcid>0</Lcid>
-      <WrapperTool>primary</WrapperTool>
-      <Isolated>False</Isolated>
-      <EmbedInteropTypes>True</EmbedInteropTypes>
-    </COMReference>
+    <ProjectReference Include="..\Impl\Impl.csproj">
+      <Project>{f89403e5-4c5c-4d8f-8401-b2d56f6d8a2b}</Project>
+      <Name>Impl</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
